pyproject.toml
[project]
name = "companionai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appnope==0.1.4",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "attrs==25.3.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.4",
    "bleach==6.2.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "comm==0.2.2",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "executing==2.2.0",
    "fastjsonschema==2.21.1",
    "fqdn==1.5.1",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "grpcio==1.73.0",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.3.0",
    "ipython-pygments-lexers==1.1.1",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "json5==0.12.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.2.5",
    "jupyter-server==2.16.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.3",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "langchain==0.3.25",
    "langchain-core==0.3.65",
    "langchain-ollama==0.3.3",
    "langchain-qdrant==0.2.0",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.8",
    "langgraph-checkpoint==2.0.26",
    "langgraph-prebuilt==0.2.2",
    "langgraph-sdk==0.1.70",
    "langsmith==0.3.45",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.3",
    "narwhals==1.42.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "notebook==7.4.3",
    "notebook-shim==0.2.4",
    "numpy==2.3.0",
    "ollama==0.5.1",
    "openai==1.86.0",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.3.0",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "portalocker==2.10.1",
    "prometheus-client==0.22.1",
    "prompt-toolkit==3.0.51",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==20.0.0",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydeck==0.9.1",
    "pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-json-logger==3.3.0",
    "python-telegram-bot==22.7",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "qdrant-client==1.14.2",
    "referencing==0.36.2",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rpds-py==0.25.1",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "send2trash==1.8.3",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.41",
    "stack-data==0.6.3",
    "streamlit==1.45.1",
    "tenacity==9.1.2",
    "terminado==0.18.1",
    "threadpoolctl==3.6.0",
    "tinycss2==1.4.0",
    "toml==0.10.2",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "types-python-dateutil==2.9.0.20250516",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uri-template==1.3.0",
    "urllib3==2.4.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "xxhash==3.5.0",
    "zstandard==0.23.0",
]
